refactor: extract collision detection to separate file
Extract is_wall() and is_blocked_cell() functions from player_movement.c
and player_slide.c into a dedicated player_collision.c file. This
consolidates collision detection logic and resolves norminette TOO_MANY_FUNCS
error in player_movement.c.
