+++ Makefile
@@ -71,6 +71,7 @@ SRCS = $(SRC_DIR)/main.c \
 	   $(SRC_DIR)/engine/raycasting/ray_draw.c \
 	   $(SRC_DIR)/engine/raycasting/texture_coords.c \
 	   $(SRC_DIR)/engine/player/player_movement.c \
+	   $(SRC_DIR)/engine/player/player_collision.c \
 	   $(SRC_DIR)/engine/player/player_rotation.c \
 	   $(SRC_DIR)/engine/player/player_slide.c \
 	   $(SRC_DIR)/engine/texture/texture_init.c \
