+++ srcs/engine/player/player_slide.c
@@ -13,7 +13,6 @@
 #include "cub3d.h"
 #include "player.h"
 
-static int	is_wall(t_game *game, double x, double y);
 static void	try_slide(t_game *game, t_slide *s);
 
 /*
@@ -56,33 +55,4 @@ static void	try_slide(t_game *game, t_slide *s)
 			&& (s->new_y - game->player.pos_y) * game->player.dir_y >= 0)
 			game->player.pos_y = s->new_y;
 	}
-}
-
-/*
-** is_wall
-** 指定位置の当たり判定をマージン付きでチェック
-** プレイヤーのバウンディングボックス四隅を確認
-*/
-static int	is_wall(t_game *game, double x, double y)
-{
-	int	x1;
-	int	x2;
-	int	y1;
-	int	y2;
-
-	x1 = (int)(x - COLLISION_MARGIN);
-	x2 = (int)(x + COLLISION_MARGIN);
-	y1 = (int)(y - COLLISION_MARGIN);
-	y2 = (int)(y + COLLISION_MARGIN);
-	if (x1 < 0 || x2 >= game->map_width || y1 < 0 || y2 >= game->map_height)
-		return (1);
-	if (game->world_map[y1][x1] != 0)
-		return (1);
-	if (game->world_map[y1][x2] != 0)
-		return (1);
-	if (game->world_map[y2][x1] != 0)
-		return (1);
-	if (game->world_map[y2][x2] != 0)
-		return (1);
-	return (0);
 }
